Fix problem with out-of-bounds array access in Map

+++ src/game/view.rs
@@ -8,6 +8,6 @@ pub struct View<T> {
     pub is_alive: bool,
     pub snakes: Vec<Snake<T>>,
     pub pellets: Vec<Pellet<T>>,
-    pub map: Vec<Vec<u8>>,
+    pub map: Vec<Vec<u32>>,
     pub self_coordinate: (usize, usize),
 }
